+++ sr_port/time_calc.c
@@ -3,7 +3,7 @@
  * Copyright (c) 2001-2018 Fidelity National Information	*
  * Services, Inc. and/or its subsidiaries. All rights reserved.	*
  *								*
- * Copyright (c) 2018-2019 YottaDB LLC and/or its subsidiaries.	*
+ * Copyright (c) 2018-2020 YottaDB LLC and/or its subsidiaries.	*
  * All rights reserved.						*
  *								*
  *	This source code contains the intellectual property	*
@@ -74,8 +74,8 @@ void	add_int_to_abs_time(ABS_TIME *atps, int4 ival, ABS_TIME *atpd)
 		 */
 		rts_error_csa(CSA_ARG(NULL) VARLSTCNT(1) ERR_TIMEROVFL);
 	}
-	ival_sec  = ival / MILLISECS_IN_SEC;					/* milliseconds -> seconds */
-	ival_nsec = (ival - (ival_sec * MILLISECS_IN_SEC)) * NANOSECS_IN_MSEC;	/* nanosecond remainder */
+	ival_sec  = ival / MILLISECS_IN_SEC;						/* milliseconds -> seconds */
+	ival_nsec = (ival - (ival_sec * MILLISECS_IN_SEC)) * (uint8)NANOSECS_IN_MSEC;	/* nanosecond remainder */
 	atpd->tv_sec = atps->tv_sec + ival_sec;
 	if ((atpd->tv_nsec = atps->tv_nsec + ival_nsec) >= NANOSECS_IN_SEC)
 	{
@@ -102,7 +102,7 @@ void add_uint8_to_abs_time(ABS_TIME *atps, uint8 ival, ABS_TIME *atpd)
 	gtm_uint8	ival_sec, ival_nsec;
 
 	ival_sec  = ival / NANOSECS_IN_SEC;			/* nanoseconds -> seconds */
-	ival_nsec = ival - (ival_sec * NANOSECS_IN_SEC);		/* nanosecond remainder */
+	ival_nsec = ival - (ival_sec * (uint8)NANOSECS_IN_SEC);		/* nanosecond remainder */
 	atpd->tv_sec = atps->tv_sec + ival_sec;
 	if ((atpd->tv_nsec = atps->tv_nsec + ival_nsec) >= NANOSECS_IN_SEC)
 	{
